Create BMEG Beginner Tutorial.md
This is a beginner tutorial for BMEG developed by Mekdes Hilete. It is in its last stages of development. I suggest making a Google Colaboratory script where a user can interact with the code used in this tutorial. 

There is a sub-tutorial that is designed to fill-in the gaps of this tutorial. It is called Graph Traversal  Sub-Tutorial. I suggest adding the url of that tutorial to the main tutorial. 

Spaces where the links of Google Colaboratory and the sub-tutorial should go are indicted in this document. 

Lastly, it would be helpful to check for errors of any sort, play around with some formatting, and make sure all the links found in this tutorial are working before this tutorial is published.